19 2 2020 New features extend the course by my own

+++ admin/users.php
@@ -1,5 +1,6 @@
 <?php include("includes/header.php"); ?>
 <?php if(!$session->is_signedIn()) redirect("login.php"); ?>
+<?php if($_SESSION['user_role'] == "subscriber") redirect("../index.php");?>
 
         <!-- Navigation -->
         <nav class="navbar navbar-inverse navbar-fixed-top" role="navigation">
@@ -31,7 +32,7 @@
                         </p>
                         
                     <div class="col-md-12">
-                        <a class="btn btn-primary" href="add_user.php">Add User</a>
+                        <a class="btn btn-primary" href="add_user.php">Add New User</a>
                             <table class="table table-hover">
                                 <thead>
                                     <tr>
@@ -40,9 +41,37 @@
                                         <th>Username</th>
                                         <th>First Name</th>
                                         <th>Second Name</th>
+                                        <th>E-mail</th>
+                                        <th>Rigistered</th>
+                                        <th>Role</th>
+                                        <th>Admin</th>
+                                        <th>Subscriber</th>
                                     </tr>
                                 </thead>
                                 
+<!-- Change user role by $_GET[''] using <a> link tags -->
+<?php 
+    
+    // in case we clicked on admin link
+    if(isset($_GET['admin'])){
+        
+        $user_id = $_GET['admin'];
+        set_user_role($user_id, "admin");
+    }
+        
+    // in case we clicked on subscriber link
+    if(isset($_GET['subscriber'])){
+        
+        $user_id = $_GET['subscriber'];
+        set_user_role($user_id, "subscriber");
+    }
+        
+                                
+?>
+                                
+                                
+              
+                                          
 <!-- Get users infos from DB -->
 <?php 
 
@@ -56,6 +85,9 @@
         $username = $user -> username;
         $firstname = $user -> first_name;
         $secondname = $user -> second_name;
+        $user_email = $user -> user_email;
+        $user_register_date = $user -> user_register_date;
+        $user_role = $user -> user_role;
         
 
         
@@ -71,7 +103,14 @@
                                         </a>
                                         
                                         <div class="pictures-links">
+                                        
+                                        <!-- Hide the loged in user delete link -->
+                                        <? if($user->user_id !== $_SESSION['user_id']): ?>
                                             <a class="delete-link"  href="delete_user.php?user_id=<? echo $id;?>">Delete</a>
+                                        <? endif; ?>   
+
+                                            
+                                            
                                             <a href="edit_user.php?user_id=<? echo $id; ?>">Edit</a>
                                             <a href="#">View</a>
                                         </div>
@@ -80,6 +119,12 @@
                                     <td> <? echo $username; ?> </td>
                                     <td> <? echo $firstname; ?> </td>
                                     <td> <? echo $secondname; ?> </td>
+                                    <td> <? echo $user_email; ?> </td>
+                                    <td> <? echo $user_register_date; ?> </td>
+                                    <td> <? echo $user_role; ?> </td>
+                                    <td><a href="users.php?admin=<? echo $id ?>">Admin</a></td>
+                                    <td><a href="users.php?subscriber=<? echo $id ?>">Subscriber</a></td>
+                                    <td>  </td>
 
                                     </tr>
 <?php endforeach; ?>
